Extract hook_reconciler from scan_controller
